+++ src/index.ts
@@ -1,6 +1,7 @@
-/// <reference path="./types/index.d.ts" />
+export * from "./types/index.js"
 
 export * from "./core/DiscordJSX.js";
-export * from "./utils/index.js";
-
 export * from "./core/hooks/useModal.js";
+
+export * from "./utils/markComponentsDisabled.js";
+export * from "./utils/resolve.js";
