+++ src/instance/Instance.ts
@@ -1,9 +1,9 @@
 import { type AttachmentPayload } from "discord.js";
 import { createRoot, Root, type InternalNode } from "../reconciler/index.js";
-import { MessageUpdater } from "../updater/index.js";
 import { PayloadBuilder, PayloadBuilderHooks } from "../payload/PayloadBuilder.js";
 import { resolveFile } from "../utils/resolve.js";
 import { createErrorPayload } from "../utils/error.js";
+import { MessageUpdater } from "../internals.js";
 
 export interface InstanceHooks extends Omit<PayloadBuilderHooks, "addAttachment"> { };
