+++ src/internals.ts
@@ -1,5 +1,12 @@
-export * from "./updater/index.js";
+export * from "./updater/MessageUpdater.js";
+export * from "./updater/update-target.js";
 export * from "./instance/Instance.js";
 export * from "./reconciler/index.js";
 export * from "./payload/PayloadBuilder.js";
 export * from "./payload/markdown.js";
+export * from "./utils/debounceAsync.js";
+export * from "./utils/error.js";
+export * from "./utils/markComponentsDisabled.js";
+export * from "./utils/mutex.js";
+export * from "./utils/resolve.js";
+export * from "./utils/timer.js";
