+++ src/types/props/content/file.ts
