+++ src/types/props/content/gallery-item.ts
