+++ src/types/props/content/separator.ts
