+++ src/types/props/content/thumbnail.ts
