+++ src/types/props/layout/label.ts
