--- src/updater/index.ts
@@ -1,3 +1,0 @@
-export * from "./MessageUpdater.js"
-export * from "./types.js"
-export * from "./utils.js"
